Rename folder to student-course-card
